requirements.txt
fastapi
uvicorn
jinja2
pillow
requests
python-multipart